refactor: 이벤트 파일 패키지 구조 재정리
