+++ hub/src/main/java/com/sparta/hub/application/event/HubEventListener.java
@@ -1,6 +1,6 @@
-package com.sparta.hub.routes.application.event;
+package com.sparta.hub.application.event;
 
-import com.sparta.hub.routes.domain.events.HubCreatedEvent;
+import com.sparta.hub.domain.events.HubCreatedEvent;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
